Rework VirtualDeice plugin

+++ plugins/virtual-device/src/commonMain/kotlin/com/alekso/dltstudio/plugins/virtualdevice/db/VirtualDeviceEntity.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class VirtualDeviceEntity(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val title: String,
     val width: Int,
     val height: Int,
